feat: Add ListItems component
This commit adds a new component `ListItems.js` that displays a list of items fetched from the Firestore database. The component uses the `@firebase/firestore` package to query the database and retrieve the items. It also includes a `DeleteItem` component for deleting individual items. This component will be used to render the list of items in the application.