chore: Add AddItem component
This commit adds a new component `AddItem.js` that allows users to add new items to the Firestore database. The component includes a form with an input field and a submit button. When the form is submitted, the value entered in the input field is added as a new document to the "items" collection in the database. This component will be used to add items to the application.