feat: Add DeleteItem component
This commit adds a new component `DeleteItem.js` that allows users to delete items from the Firestore database. The component includes a button that, when clicked, deletes the corresponding item document from the "items" collection in the database. This component will be used to provide the delete functionality in the application.